Add initial interactivity mode supporting word deletes
Add -interactive flag that prompts the user for each word to confirm
it was accepted by the game before continuing. If it was not accepted
the user is prompted to allow removal of word from the dictionary. If
the user confirms deletion the word is deleted from the word file and
the word file is rewritten.